wip(cli): more modernization
- Read known list of ACT rules directly from @qualweb/act-rules instead of local list
- Add more enums, segregate into their own file
- Improve documentation
- Add output options (not yet implemented)
