--- packages/cli/src/lib/act-rules.ids.ts
@@ -1,151 +1,0 @@
-export const actRulesIds = {
-  qualweb_id: [
-    'QW-ACT-R1',
-    'QW-ACT-R2',
-    'QW-ACT-R3',
-    'QW-ACT-R4',
-    'QW-ACT-R5',
-    'QW-ACT-R6',
-    'QW-ACT-R7',
-    'QW-ACT-R9',
-    'QW-ACT-R10',
-    'QW-ACT-R11',
-    'QW-ACT-R12',
-    'QW-ACT-R13',
-    'QW-ACT-R14',
-    'QW-ACT-R15',
-    'QW-ACT-R16',
-    'QW-ACT-R17',
-    'QW-ACT-R18',
-    'QW-ACT-R19',
-    'QW-ACT-R20',
-    'QW-ACT-R21',
-    'QW-ACT-R22',
-    'QW-ACT-R23',
-    'QW-ACT-R24',
-    'QW-ACT-R25',
-    'QW-ACT-R26',
-    'QW-ACT-R27',
-    'QW-ACT-R28',
-    'QW-ACT-R29',
-    'QW-ACT-R30',
-    'QW-ACT-R31',
-    'QW-ACT-R32',
-    'QW-ACT-R33',
-    'QW-ACT-R34',
-    'QW-ACT-R35',
-    'QW-ACT-R36',
-    'QW-ACT-R37',
-    'QW-ACT-R38',
-    'QW-ACT-R39',
-    'QW-ACT-R40',
-    'QW-ACT-R41',
-    'QW-ACT-R42',
-    'QW-ACT-R43',
-    'QW-ACT-R44',
-    'QW-ACT-R48',
-    'QW-ACT-R49',
-    'QW-ACT-R50',
-    'QW-ACT-R51',
-    'QW-ACT-R52',
-    'QW-ACT-R53',
-    'QW-ACT-R54',
-    'QW-ACT-R56',
-    'QW-ACT-R57',
-    'QW-ACT-R58',
-    'QW-ACT-R59',
-    'QW-ACT-R60',
-    'QW-ACT-R61',
-    'QW-ACT-R62',
-    'QW-ACT-R63',
-    'QW-ACT-R64',
-    'QW-ACT-R65',
-    'QW-ACT-R66',
-    'QW-ACT-R67',
-    'QW-ACT-R68',
-    'QW-ACT-R69',
-    'QW-ACT-R70',
-    'QW-ACT-R71',
-    'QW-ACT-R72',
-    'QW-ACT-R73',
-    'QW-ACT-R74',
-    'QW-ACT-R75',
-    'QW-ACT-R76',
-    'QW-ACT-R77'
-  ],
-  act_id: [
-    '2779a5',
-    'b5c3f8',
-    '5b7ae0',
-    'bc659a',
-    'bf051a',
-    '59796f',
-    'b33eff',
-    'b20e66',
-    '4b1c6c',
-    '97a4e1',
-    'c487ae',
-    '6cfa84',
-    'b4f0c3',
-    '80f0bf',
-    'e086e5',
-    '23a2a8',
-    '3ea0c8',
-    'cae760',
-    '674b10',
-    '7d6734',
-    'de46e4',
-    'c5a4ea',
-    '73f2c2',
-    '5c01ea',
-    'eac66b',
-    '5f99a7',
-    '4e8ab6',
-    'e7aa44',
-    '2ee8b8',
-    'c3232f',
-    '1ec09b',
-    'ff89c9',
-    '6a7281',
-    'ffd0e9',
-    'a25f45',
-    'afw4f7',
-    'bc4a75',
-    'd0f69e',
-    '59br37',
-    '36b590',
-    '8fc3b6',
-    '0ssw9k',
-    'fd3a94',
-    '46ca7f',
-    'aaa1bf',
-    '4c31df',
-    'fd26cf',
-    'ac7dc6',
-    'ee13b5',
-    'd7ba54',
-    '1ea59c',
-    'ab4d13',
-    'f196ce',
-    '2eb176',
-    'afb423',
-    'f51b46',
-    '1a02b0',
-    'oj04fd',
-    'b40fd1',
-    '047fe0',
-    '307n5z',
-    'm6b1q3',
-    '24afc2',
-    '78fd32',
-    '9e45ec',
-    'akn7bn',
-    'bisz58',
-    '8a213c',
-    '3e12e1',
-    'ye5d6e',
-    'cf77f2',
-    '09o5cg',
-    'in6db8'
-  ]
-};
