+++ packages/cli/src/lib/types.ts
@@ -1,0 +1,23 @@
+/**
+ * Option values for --module.
+ * Kept here in one place to avoid repeated strings.
+ */
+export enum ModuleOptionsEnum {
+  ACTRules = 'act-rules',
+  BestPractices = 'best-practices',
+  WCAGTechniques = 'wcag-techniques',
+  Counter = 'counter',
+};
+
+export enum ConformanceLevelEnum {
+  A = 'A',
+  AA = 'AA',
+  AAA = 'AAA',
+};
+
+export enum PrincipleEnum {
+  Perceivable = 'Perceivable',
+  Operable = 'Operable',
+  Understandable = 'Understandable',
+  Robust = 'Robust',
+};
