+++ packages/cli/src/modern.ts
@@ -3,132 +3,11 @@
  */
 import { Command, Option, OptionValues } from 'commander';
 
-import actRulesJson from '@qualweb/act-rules/lib/rules.json';
 import { ACTRules } from '@qualweb/act-rules';
-import { ModuleOptions, QualWeb, QualwebOptions } from '@qualweb/core';
-const allActRules = Object.values(actRulesJson);
-
-/**
- * Option values for --module.
- * Kept here in one place to avoid repeated strings.
- */
-enum ModuleOptionsEnum {
-  ACTRules = 'act-rules',
-  BestPractices = 'best-practices',
-  WCAGTechniques = 'wcag-techniques',
-  Counter = 'counter',
-};
-
-type ActRuleListParseHelperAccumulator = {
-  ok: string[],
-  error: string[],
-};
-
-/**
- * Accumulator helper function for parsing ACT rules. Pass to
- * {@link Option.argParser}. Gathers all inputs in an object with two arrays,
- * one for valid rules ("ok")  and one for invalid rules ("error"). * 
- * @param value Current input value. CommanderJS calls this function once for
- * each value that was passed.
- * @param previousValue Accumulator value.
- * @returns 
- */
-function actRulesListParseHelper(value: string, previousValue: ActRuleListParseHelperAccumulator): ActRuleListParseHelperAccumulator {
-  // Initialize result array if undefined.
-  if (!previousValue)
-    previousValue = { error: [], ok: [] };
-
-  if (value.match(/^QW-ACT-R\d+$/i)) {
-    // parse as QualWeb internal name.
-
-    // QW codes are all upper case.
-    value = value.toUpperCase();
-
-    const foundActRule = allActRules.find(rule => rule.code === value);
-
-    if (foundActRule) {
-      previousValue.ok.push(foundActRule.code);
-    } else {
-      previousValue.error.push(value);
-    }
-  } else if (value.length === 6) {
-    // Parse as official ACT rule hash.
-
-    // ACT hashes are all lower case.
-    value = value.toLowerCase();
-
-    const foundActRule = allActRules.find(rule => rule.mapping === value);
-
-    if (foundActRule) {
-      previousValue.ok.push(foundActRule.code);
-    } else {
-      previousValue.error.push(value);
-    }
-  } else {
-    previousValue.error.push(value);
-  }
-  
-  return previousValue;
-}
-
-/**
- * Adds input options (URL, file, crawl) to a command.
- * @param command The command to add the options to. This *will* modify the
- * command object.
- * @returns The modified command object. Good for chaining.
- */
-function addInputOptionsToCommand(command: Command): Command {
-  const urlInputOption = new Option('-u, --url <url>', 'URL to test');
-  const fileInputOption = new Option('-f, --file <file>', 'File to test');
-  const crawlInputOption = new Option('-c, --crawl <crawl>', 'Crawl a website');
-
-  urlInputOption.conflicts([fileInputOption.attributeName(), crawlInputOption.attributeName()]);
-  fileInputOption.conflicts([urlInputOption.attributeName(), crawlInputOption.attributeName()]);
-  crawlInputOption.conflicts([urlInputOption.attributeName(), fileInputOption.attributeName()]);
-
-  command.addOption(urlInputOption);
-  command.addOption(fileInputOption);
-  command.addOption(crawlInputOption);
-
-  return command;
-}
-
-/**
- * Adds ACT rule options (include/exclude/level filter) to a command.
- * @param command The command to add the options to. This *will* modify the
- * command object.
- * @returns The modified command object. Good for chaining.
- */
-function addActRuleOptionsToCommand(command: Command): Command {
-  const actRuleIncludeOption = new Option('--act-rules <rules...>', 'ACT rules to include')
-    .argParser(actRulesListParseHelper)
-    .implies({ module: 'act-rules' })
-    ;
-
-  const actRuleExcludeOption = new Option('--exclude-act <rules...>', 'ACT rules to include')
-    .argParser(actRulesListParseHelper)
-    .implies({ module: 'act-rules' })
-    ;
-
-  const actRuleLevelOption = new Option('--act-levels <levels...>', 'ACT level to test')
-    .choices(['A', 'AA', 'AAA'])
-    .implies({ module: 'act-rules' })
-    ;
-
-  // Allowing users to include AND exclude specific rules doesn't make sense.
-  // It's either ONLY include a few rules or ALL rules EXCEPT a few.
-  actRuleIncludeOption.conflicts(actRuleExcludeOption.attributeName());
-  actRuleExcludeOption.conflicts(actRuleIncludeOption.attributeName());
-  // For level, it doesn't make sense to use the include filter, as that's
-  // implied. Excluding specific rules despite the level filter is fine.
-  actRuleLevelOption.conflicts([actRuleIncludeOption.attributeName()]);
-
-  command.addOption(actRuleIncludeOption);
-  command.addOption(actRuleExcludeOption);
-  command.addOption(actRuleLevelOption);
-
-  return command;
-}
+import { QualwebOptions } from '@qualweb/core';
+import { addActRuleOptionsToCommand, addInputOptionsToCommand, addListActRulesSubcommand } from './lib/actRules';
+import { ModuleOptionsEnum } from './lib/types';
+import { addOutputOptions } from './lib/outputOptions';
 
 /**
  * Checks that the ACT rule options parsed from user input is valid. Prints
@@ -194,17 +73,24 @@ async function main(): Promise<void> {
   const moduleToRunOption = new Option('-m, --module <modules...>', 'Modules to run')
     // Use enum values for choices to avoid typos and string repetition.
     .choices(Object.values(ModuleOptionsEnum))
+    .default([])
     ;
-
   program.addOption(moduleToRunOption);
 
+  /** GENERAL options */
+  addOutputOptions(program);
+
+  /** ACT options */
   addActRuleOptionsToCommand(program);
+
+  // Add list-act-rules subcommand to dump known ACT rules to output.
+  addListActRulesSubcommand(program);
 
   program.parse();
 
   const opts = program.opts();
 
-  opts.modules.forEach((module: ModuleOptionsEnum) => {
+  opts.module.forEach((module: ModuleOptionsEnum) => {
     modulesToRun.add(module);
   });
